LPS-87059 Move applicable bean component classes to internal package in portal-template-impl module
