+++ modules/apps/portal-template/portal-template-impl/src/main/java/com/liferay/portal/template/internal/DDMTemplateResourceParser.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.template;
+package com.liferay.portal.template.internal;
 
 import com.liferay.dynamic.data.mapping.kernel.DDMStructureManagerUtil;
 import com.liferay.dynamic.data.mapping.kernel.DDMTemplate;
