+++ modules/apps/portal-template/portal-template-impl/src/main/java/com/liferay/portal/template/internal/ThemeResourceParser.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.template;
+package com.liferay.portal.template.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
